+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.2.0](https://github.com/euberdeveloper/ga-project-version/compare/v1.3.0...v2.2.0) (2023-08-17)
+
+
+### Features
+
+* add poetry and pipfile ([c2c8dfb](https://github.com/euberdeveloper/ga-project-version/commit/c2c8dfb226d709233d2ea90c93fab495e411cfb4))
+
+
+### Bug Fixes
+
+* **source:** fix get version only string ([382b0ba](https://github.com/euberdeveloper/ga-project-version/commit/382b0baf6d6eb2501b9a0f62cf68d28df89a898e))
+
 ## [2.1.0](https://github.com/euberdeveloper/ga-project-version/compare/v1.3.0...v2.1.0) (2023-03-11)